requirements.txt
langchain==0.3.27
langchain_openai==0.3.30
python-dotenv==1.1.1
Requests==2.32.4
streamlit==1.48.1
